+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.6.0
+* `name` in Case is deprecated. It was the same as the key in Uibook Controller,
+  so it was unnecessary duplication. You should set Page Name in
+  the Uibook Controller now.
+
 ## 0.5.8
 * Added support of Cyrillic component names
 * Fixed converting of CamelCase to space separated words in Page Selector